+++ src/tests/Stream/MessageWriterTests.cpp
@@ -9,7 +9,9 @@
 
 #include "Metrics/Registrar.h"
 #include "Stream/MessageWriter.h"
+#include "WriterModule/ep00/ep00_Writer.h"
 #include "WriterModuleBase.h"
+#include "WriterRegistrar.h"
 #include "helpers/SetExtractorModule.h"
 #include <array>
 #include <gtest/gtest.h>
@@ -18,7 +20,8 @@
 class WriterModuleStandIn : public WriterModule::Base {
 public:
   MAKE_MOCK0(config_post_processing, void(), override);
-  WriterModuleStandIn() : WriterModule::Base(true, "test") {}
+  WriterModuleStandIn(std::vector<std::string> const &ExtraModules = {})
+      : WriterModule::Base(true, "test", ExtraModules) {}
   ~WriterModuleStandIn() = default;
   MAKE_CONST_MOCK1(init_hdf, WriterModule::InitResult(hdf5::node::Group &),
                    override);
@@ -28,11 +31,54 @@ public:
 
 class DataMessageWriterTest : public ::testing::Test {
 public:
+  void SetUp() override {
+    WriterModule::Registry::clear();
+    WriterModule::Registry::Registrar<WriterModule::ep00::ep00_Writer>
+        RegisterIt1("ep00", "epics_con_status");
+  }
   WriterModuleStandIn WriterModule;
   Metrics::Registrar MetReg{"some_prefix", {}};
 };
 
 using trompeloeil::_;
+
+TEST_F(DataMessageWriterTest, NoExtraModules) {
+  EXPECT_TRUE(WriterModule.getEnabledExtraModules().empty());
+}
+
+TEST_F(DataMessageWriterTest, InvalidExtraModule) {
+  WriterModuleStandIn TestWriterModule({"no_such_module"});
+  EXPECT_FALSE(TestWriterModule.hasExtraModules());
+  EXPECT_TRUE(TestWriterModule.getEnabledExtraModules().empty());
+}
+
+TEST_F(DataMessageWriterTest, ValidExtraModule) {
+  WriterModuleStandIn TestWriterModule({"ep00"});
+  EXPECT_TRUE(TestWriterModule.hasExtraModules());
+  EXPECT_EQ(TestWriterModule.getEnabledExtraModules().size(), 1u);
+}
+
+TEST_F(DataMessageWriterTest, ValidExtraModuleByName) {
+  WriterModuleStandIn TestWriterModule({"epics_con_status"});
+  EXPECT_TRUE(TestWriterModule.hasExtraModules());
+  EXPECT_EQ(TestWriterModule.getEnabledExtraModules().size(), 1u);
+}
+
+TEST_F(DataMessageWriterTest, DisabledExtraModule) {
+  WriterModuleStandIn TestWriterModule({"epics_con_status"});
+  REQUIRE_CALL(TestWriterModule, config_post_processing()).TIMES(1);
+  TestWriterModule.parse_config("{\"enable_epics_con_status\": false}");
+  EXPECT_TRUE(TestWriterModule.hasExtraModules());
+  EXPECT_TRUE(TestWriterModule.getEnabledExtraModules().empty());
+}
+
+TEST_F(DataMessageWriterTest, EnableExtraModule) {
+  WriterModuleStandIn TestWriterModule({"epics_con_status"});
+  REQUIRE_CALL(TestWriterModule, config_post_processing()).TIMES(1);
+  TestWriterModule.parse_config("{\"enable_epics_con_status\": true}");
+  EXPECT_TRUE(TestWriterModule.hasExtraModules());
+  EXPECT_EQ(TestWriterModule.getEnabledExtraModules().size(), 1u);
+}
 
 TEST_F(DataMessageWriterTest, WriteMessageSuccess) {
   REQUIRE_CALL(WriterModule, write(_)).TIMES(1);